[models] Added AbstractNotification #202
Implements and closes #102
