+++ setup.py
@@ -29,7 +29,8 @@ setup(
         'django>=1.7',
         'django-model-utils>=2.0.3',
         'jsonfield>=1.0.3',
-        'pytz'
+        'pytz',
+        'swapper'
     ],
     test_requires=[
         'django>=1.7',
